// src/segment-types.ts
export interface BaseSegment {
  type: string;
  /** Indicates this segment has a definition built by the SegmentBuilder. */
  definitionKey?: string;
  negate?: boolean;
  options?: Record<string, unknown>;
}
// ...
export interface ComposedSegment<TSegment extends BaseSegment, TCompositionOperator extends string = string>
  extends BaseSegment {
  type: 'composed';
  options: {
    operator: TCompositionOperator;
    segments: (ComposedSegment<TSegment, TCompositionOperator> | TSegment)[];
  };
}